Restore filenames
