+++ sdk/eventhub/azure-eventhubs/azure/eventhub/__init__.py
@@ -6,12 +6,12 @@
 __path__ = __import__('pkgutil').extend_path(__path__, __name__)  # type: ignore
 __version__ = "5.0.0b5"
 from uamqp import constants  # type: ignore
-from ._common import EventData, EventDataBatch, EventPosition
-from ._error import EventHubError, EventDataError, ConnectError, \
+from .common import EventData, EventDataBatch, EventPosition
+from .error import EventHubError, EventDataError, ConnectError, \
     AuthenticationError, EventDataSendError, ConnectionLostError
 from ._producer_client import EventHubProducerClient
 from ._consumer_client import EventHubConsumerClient
-from ._common import EventHubSharedKeyCredential, EventHubSASTokenCredential
+from .common import EventHubSharedKeyCredential, EventHubSASTokenCredential
 from ._eventprocessor.partition_manager import PartitionManager, OwnershipLostError
 from ._eventprocessor.event_processor import CloseReason
