+++ sdk/eventhub/azure-eventhubs/azure/eventhub/_connection_manager.py
@@ -83,5 +83,4 @@ def get_connection_manager(**kwargs):
     connection_mode = kwargs.get("connection_mode", _ConnectionMode.SeparateConnection)
     if connection_mode == _ConnectionMode.ShareConnection:
         return _SharedConnectionManager(**kwargs)
-    else:
-        return _SeparateConnectionManager(**kwargs)
+    return _SeparateConnectionManager(**kwargs)
