+++ sdk/eventhub/azure-eventhubs/azure/eventhub/aio/_consumer_producer_mixin_async.py
@@ -7,7 +7,7 @@ import logging
 import time
 
 from uamqp import errors, constants, compat  # type: ignore
-from .._error import EventHubError
+from ..error import EventHubError
 from ..aio.error_async import _handle_exception
 
 log = logging.getLogger(__name__)
