+++ sdk/eventhub/azure-eventhubs/azure/eventhub/aio/producer_async.py
@@ -14,9 +14,9 @@ from uamqp import SendClientAsync  # type: ignore
 from azure.core.tracing import SpanKind, AbstractSpan  # type: ignore
 from azure.core.settings import settings  # type: ignore
 
-from .._common import EventData, EventDataBatch
-from .._error import _error_handler, OperationTimeoutError, EventDataError
-from .._producer import _error, _set_partition_key, _set_trace_message
+from ..common import EventData, EventDataBatch
+from ..error import _error_handler, OperationTimeoutError, EventDataError
+from ..producer import _error, _set_partition_key, _set_trace_message
 from ._consumer_producer_mixin_async import ConsumerProducerMixin
 
 log = logging.getLogger(__name__)
