+++ sdk/eventhub/azure-eventhubs/azure/eventhub/client.py
@@ -17,13 +17,13 @@ from uamqp import Message  # type: ignore
 from uamqp import authentication  # type: ignore
 from uamqp import constants  # type: ignore
 
-from ._producer import EventHubProducer
-from ._consumer import EventHubConsumer
-from ._common import parse_sas_token, EventPosition
-from ._client_abstract import EventHubClientAbstract
-from ._common import EventHubSASTokenCredential, EventHubSharedKeyCredential
+from .producer import EventHubProducer
+from .consumer import EventHubConsumer
+from .common import parse_sas_token, EventPosition
+from .client_abstract import EventHubClientAbstract
+from .common import EventHubSASTokenCredential, EventHubSharedKeyCredential
 from ._connection_manager import get_connection_manager
-from ._error import _handle_exception
+from .error import _handle_exception
 
 if TYPE_CHECKING:
     from azure.core.credentials import TokenCredential  # type: ignore
