+++ sdk/eventhub/azure-eventhubs/azure/eventhub/configuration.py
