+++ sdk/eventhub/azure-eventhubs/azure/eventhub/error.py
